Move and rename add / clone cell helpers
